% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_module.R
\name{add_module}
\alias{add_module}
\title{Create a module}
\usage{
add_module(
  name,
  inherit = "TidyModule",
  pkg = getwd(),
  open = TRUE,
  dir_create = TRUE,
  export = FALSE
)
}
\arguments{
\item{name}{The class name of the module.}

\item{inherit}{Parent module class. Default is TidyModule.}

\item{pkg}{Path to the root of the package. Default is \code{getwd()}.}

\item{open}{Should the file be opened?}

\item{dir_create}{Creates the directory if it doesn't exist, default is \code{TRUE}.}

\item{export}{Logical. Should the module be exported? Default is \code{FALSE}.}
}
\description{
This function creates a \code{{tm}} module class inside the \verb{R/} folder.
}
\note{
As a convention, this function will automatically capitalize the first character of the \code{name} argument.
}
