% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{add_tm_snippets}
\alias{add_tm_snippets}
\title{Add \code{{tm}} snippets to RStudio}
\usage{
add_tm_snippets(force = FALSE)
}
\arguments{
\item{force}{Force the re-installation when the snippets are already installed.}
}
\description{
This function adds useful \code{{tm}} code snippets to RStudio.
}
