% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{callModules}
\alias{callModules}
\title{Retrieve module from ModStore}
\usage{
callModules()
}
\description{
This utility function retrieve tidymodules from the central ModStore
using module namespace/id and/or group
}
