% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{check_and_load}
\alias{check_and_load}
\title{check if list of package namespaces exist, load them or display relevant information}
\usage{
check_and_load(packages)
}
\arguments{
\item{p}{character verctor of package names}
}
\description{
Utility function for managing package dependencies for tidymodules examples
}
\examples{

check_and_load("ggplot2")

}
