% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{combine_ports}
\alias{combine_ports}
\title{Combine ports function}
\usage{
combine_ports(...)
}
\arguments{
\item{...}{key/value pairs of ports}
}
\description{
This function combines ports into a reactive list (reactiveValues)
}
\examples{
\dontrun{
# Somewhere in the app...
MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
MyModule$new("Mod1")
MyModule$new("Mod2")
MyModule$new("Mod3")

# Must be in the server code and after calling the modules!
callModules()
observe({
  combine_ports(
  input_1 = mod(1)$getOutput(1),
  input_2 = mod(2)$getOutput(1)
  ) \%>1\% mod(3)
})
}

}
