% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{connect_ports}
\alias{connect_ports}
\title{connect ports from two modules}
\usage{
connect_ports(leftModule = NULL, leftPort = 1, rightModule = NULL,
  rightPort = 1, reverse = FALSE)
}
\description{
connect ports from two modules
}
