% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{defineEdges}
\alias{defineEdges}
\title{Function wrapper for ports connection expression.}
\usage{
defineEdges(x)
}
\arguments{
\item{x}{expression}
}
\description{
Used in server functions to define how modules are connected to each other.
}
