% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_R6CG_list}
\alias{get_R6CG_list}
\title{Recursive function for retrieving R6ClassGenerator inheritance}
\usage{
get_R6CG_list(r6cg)
}
\arguments{
\item{r6cg}{A R6ClassGenerator object.}
}
\value{
vector of class names
}
\description{
This function is used to retrieve a list of class name that a R6ClassGenerator object inherit from.
}
