% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMod}
\alias{getMod}
\title{Retrieve module from ModStore}
\usage{
getMod(id = 1, group = NULL)
}
\description{
This utility function retrieve tidymodules from the central ModStore
using module namespace/id and/or group
}
