% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getSessionId}
\alias{getSessionId}
\title{Function that generates session Id}
\usage{
getSessionId(out = NULL)
}
\arguments{
\item{out}{A shiny output as provide by the shiny server function.}
}
\value{
A session ID
}
\description{
tidymodules offers the ability to manage application sessions.
This function is the main function used by tidymodules to find the current session Id.
It takes an optional shiny out as argument. If null, default to the global_session.
}
