% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%:c>:\%}
\alias{\%:c>:\%}
\title{Multi-port mapping function}
\usage{
l \%:c>:\% r
}
\arguments{
\item{l}{left module.}

\item{r}{right module.}
}
\value{
The right module
}
\description{
This pipe copy all the left output ports to the right input ports.
This is a copy not a mapping. The right module input ports will be created.
}
