% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%:>:\%}
\alias{\%:>:\%}
\title{Multi-port mapping function}
\usage{
l \%:>:\% r
}
\arguments{
\item{l}{left module}

\item{r}{right module}
}
\value{
The right module
}
\description{
This pipe maps all the left output ports to the right input ports.
}
