% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%:i:\%}
\alias{\%:i:\%}
\title{Multi-port mapping function}
\usage{
l \%:i:\% r
}
\arguments{
\item{l}{left module.}

\item{r}{right module.}
}
\value{
The right module.
}
\description{
This pipe maps all the left input ports to the right input ports.
}
