% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%:pi:\%}
\alias{\%:pi:\%}
\title{Multi-port mapping function}
\usage{
l \%:pi:\% r
}
\arguments{
\item{l}{parent module.}

\item{r}{child module.}
}
\value{
The child module.
}
\description{
This pipe copy all the left input ports to the right input ports in a parent to child mode.
}
