% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%->\%}
\alias{\%->\%}
\title{Port mapping function (port level)}
\usage{
lp \%->\% rp
}
\value{
The module of the right port
}
\description{
This pipe works at the port level where left and right object are ports not modules.
Take the left port and maps it to the right port.
}
