% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%->>\%}
\alias{\%->>\%}
\title{Port mapping function (port level) - forward version}
\usage{
lp \%->>\% rp
}
\arguments{
\item{lp}{Left port.}

\item{rp}{Right port.}
}
\value{
The left port
}
\description{
This pipe works at the port level where left and right object are ports not modules.
Take the left port and maps it to the right port. This is a forward version, i.e. return the left item.
}
