% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%>>y\%}
\alias{\%>>y\%}
\alias{\%>>1\%}
\alias{\%>>2\%}
\alias{\%>>3\%}
\alias{\%>>4\%}
\alias{\%>>5\%}
\alias{\%>>6\%}
\alias{\%>>7\%}
\alias{\%>>8\%}
\alias{\%>>9\%}
\alias{\%>>10\%}
\title{Input port mapping function}
\description{
This pipe maps the left object (must be a reactive function
or a reactivevalues object) to the right module's input port defined by
the number in the operator (y).
}
