% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%x<<y\%}
\alias{\%x<<y\%}
\title{Single-port mapping function (Reverse version)}
\value{
The right module
}
\description{
This pipe works at the module level.
It maps the right module's output port defined by the right number (y) of the pipe operator to the left module's input port defined by the left number (x).
}
