% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{iport}
\alias{iport}
\title{Retrieve input module's port}
\usage{
iport(id = 1, p = 1, g = NULL)
}
\arguments{
\item{id}{Name or Id of the module}

\item{p}{Port Id or name}

\item{g}{Module group name}
}
\description{
This utility function retrieve the tidymodules input port specified in the arguments.
}
