% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{listModules}
\alias{listModules}
\title{List modules in current session}
\usage{
listModules(verbose = FALSE, global = FALSE)
}
\arguments{
\item{verbose}{Display module description as well}

\item{global}{use the global session? Default to FALSE}
}
\description{
This function list module objects found in the current session
}
