% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{make_double_pipe}
\alias{make_double_pipe}
\title{make_double_pipe: Pipe function generator}
\usage{
make_double_pipe(l, r, f = FALSE, rev = FALSE)
}
\arguments{
\item{l}{Left module.}

\item{r}{Right module.}

\item{f}{fast boolean. Used with the '>>' or '<<' signs to return the item at the opposite of the arrow.}

\item{rev}{Reverse operation. Boolean that indicates if this is a reverse operation, i.e. '<' or '<<'.}
}
\description{
Create a pipe function for mapping a module output to a module input
}
\keyword{internal}
