% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{make_single_pipe}
\alias{make_single_pipe}
\title{make_single_pipe: Pipe function generator}
\usage{
make_single_pipe(p = NULL, f = FALSE, rev = FALSE)
}
\arguments{
\item{f}{fast boolean. Used with the '>>' or '<<' signs to return the item at the opposite of the arrow.}

\item{rev}{Reverse operation. Boolean that indicates if this is a reverse operation, i.e. '<' or '<<'.}
}
\description{
Create a pipe function for mapping a reactive expression/value to a module input
}
\keyword{internal}
