% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{map_ports}
\alias{map_ports}
\title{Ports mapping function}
\usage{
map_ports(
  leftModule = NULL,
  leftPort = 1,
  rightModule = NULL,
  rightPort = 1,
  reverse = FALSE
)
}
\arguments{
\item{leftModule}{The left module object}

\item{leftPort}{Port name or Id of the left module's output port}

\item{rightModule}{The right module object}

\item{rightPort}{Port name or Id of the right module's input port}

\item{reverse}{ligical value indicating which module to return. Default to FALSE, the right module}
}
\description{
This function maps a module's outpout port to another module's input port.
}
\details{
connect ports from two different modules
}
