% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{mkDoublePipe}
\alias{mkDoublePipe}
\title{mkDoublePipe: Pipe function generator}
\usage{
mkDoublePipe(l, r, f = TRUE, rev = FALSE)
}
\description{
Create a pipe function for mapping module output to module input
}
\keyword{internal}
