% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModStore.R
\docType{class}
\name{ModStore}
\alias{ModStore}
\title{Create a R6 ModStore class}
\format{\code{\link{R6Class}} object.}
\usage{
ModStore
}
\value{
Object of \code{\link{R6Class}} with methods for creating and managing tidy modules.
}
\description{
This class is used to create a state container for tidy modules
}
\section{Fields}{

\describe{
\item{\code{dbName}}{Path to SQLight DB file. Default to getOption("dbName") or ./.log/log.sqlite
Note that the DB file as well as its parent folder (.log) need to be writable by others 
to properly collect usage logs.}

\item{\code{identity}}{Authentication object \code{\link{Authentication}}}

\item{\code{force}}{Force tracking on localhost, default to false.}
}}

\section{Methods}{


\describe{
  \item{\code{new(identity)}}{This method is used to create object of this class with \code{identity} as Authentication object.}
  \item{\code{log2db(action,param,desc)}}{This method log the specific action described with the variables action/param/desc into the tracking database.}
  \item{\code{getDbName()}}{This method return the path to the SQLite databasse file.}
  }
}

\examples{
rvalues <- tidymodules::init(input,output,session)
observe({
  
  rvalues$SessionTracking$log2db(
    action="access",
    param="home",
    desc=getOption("deployVersion")
  )
})
}
\keyword{datasets}
