% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{multiPipeFunc}
\alias{multiPipeFunc}
\title{multiPipeFunc}
\usage{
multiPipeFunc(l_mod, r_mod, rev = FALSE, t = NULL)
}
\arguments{
\item{l_mod}{Left module.}

\item{r_mod}{Right module.}

\item{rev}{Reverse operation. Boolean.}

\item{t}{TO DO.}
}
\description{
Pipe function for sequentially mapping left module outputs to
right module inputs
}
\keyword{internal}
