% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{oport}
\alias{oport}
\title{Retrieve output module's port}
\usage{
oport(id = 1, p = 1, g = NULL)
}
\arguments{
\item{id}{Name or Id of the module}

\item{p}{Port Id or name}

\item{g}{Module group name}
}
\description{
This utility function retrieve the tidymodules output port specified in the arguments.
}
