% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%->\%}
\alias{\%->\%}
\alias{\%:>:\%}
\alias{\%:i:\%}
\alias{\%:>>:\%}
\title{Function for directly mapping output port / reactive to input port}
\usage{
lp \%->\% rp

l \%:>:\% r

l \%:i:\% r

l \%:>>:\% r
}
\description{
Function for directly mapping output port / reactive to input port

Forward output/input multi-port pipe

Forward input/input multi-port pipe

#' Forward input/input multi-port pipe
#' 
#' @rdname pipes
#' @export %:o:%
"%:o:%" <- function(l,r) { return(multiPipeFunc(l,r,t="output")); }
Forward output/input multi-port pipe that returns the left module
}
