% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{port}
\alias{port}
\title{Retrieve module's port}
\usage{
port(id = 1, p = 1, t = "in", g = NULL)
}
\arguments{
\item{id}{Name or Id of the module}

\item{p}{Port Id or name}

\item{t}{Port type, in or out}

\item{g}{Module group name}
}
\description{
This utility function retrieve the tidymodules port specified in the arguments.
}
