% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{port_map}
\alias{port_map}
\title{port_map}
\usage{
port_map(lp, rp, f = FALSE)
}
\arguments{
\item{lp}{Left port}

\item{rp}{Right port}

\item{f}{Forward operartion. Boolean.}
}
\description{
port mapping function.
}
\keyword{internal}
