% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{race_ports}
\alias{race_ports}
\title{collapse ports into a single port and make them race
i.e. always return the last one updated}
\usage{
race_ports(racers = NULL, to = NULL, port = NULL)
}
\description{
collapse ports into a single port and make them race
i.e. always return the last one updated
}
