% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{session_type}
\alias{session_type}
\title{List of possible session types}
\format{
An object of class \code{list} of length 3.
}
\usage{
session_type
}
\description{
tidymodules offers the ability to manage application sessions.
At the moment the three options below are available.

\itemize{

\item{SHINY}{ : The default behaviour of shiny application and the default for tidymodules. Every time you access an application
you get a new token Id that defines your application user session.}

\item{USER}{ : This method defines a session based on the information available in the request object of shiny output.
It is a concatenation of the variables REMOTE_ADDR, HTTP_HOST and PATH_INFO like below.

\code{sid <- paste0(r$REMOTE_ADDR,"@",r$HTTP_HOST,r$PATH_INFO))}

Note that the method is actually not working properly for now as the information available via the request object
are not reflecting the actual user. We are working on a better method to uniquely identify a remote user.}

\item{CUSTOM}{ : This method allow the developper to provide a custom function for generating the session Id.
It relies on the global options \code{tm_session_custom} being set and pointing to a function taking a shiny output as argument.}

}
}
\keyword{datasets}
