% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{showExamples}
\alias{showExamples}
\title{Launcher for the tidymodules examples}
\usage{
showExamples(id = NULL, server = F, options = NULL)
}
\arguments{
\item{id}{Example ID. If null display list of examples with ID.}

\item{server}{boolean. Is this a server call?}

\item{options}{list of options to be passed to shinyApps or shinyDir}
}
\description{
Helper function to launch the tidymodules examples.
}
\examples{

if (interactive()) {

 showExamples(1)

}
}
