% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Store.R
\docType{class}
\name{Store}
\alias{Store}
\title{Module to explore content of ModStore}
\format{\code{\link{R6Class}} object.}
\usage{
Store
}
\value{
Object of \code{\link{R6Class}}
}
\description{
List applications / sessions / modules / edges in current process
}
\keyword{datasets}
