% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyModule.R
\docType{class}
\name{TidyModule}
\alias{TidyModule}
\title{Create a R6 Module class}
\format{\code{\link{R6Class}} object.}
\usage{
TidyModule
}
\value{
Object of \code{\link{R6Class}} with methods for creating and managing tidymodules modules.
}
\description{
This is the main class provided by tidymodules and should be the parent of all tidymodules modules.
It provides methods for defining input and output communication ports, assigning them and managing module namespace.
}
\section{Fields}{

\describe{
\item{\code{dbName}}{Path to SQLight DB file. Default to getOption("dbName") or ./.log/log.sqlite
Note that the DB file as well as its parent folder (.log) need to be writable by others 
to properly collect usage logs.}

\item{\code{identity}}{Authentication object \code{\link{Authentication}}}

\item{\code{force}}{Force tracking on localhost, default to false.}
}}

\section{Methods}{


\describe{
  \item{\code{new(identity)}}{This method is used to create object of this class with \code{identity} as Authentication object.}
  \item{\code{log2db(action,param,desc)}}{This method log the specific action described with the variables action/param/desc into the tracking database.}
  \item{\code{getDbName()}}{This method return the path to the SQLite database file.}
  }
}

\examples{
rvalues <- ava::init(input,output,session)
observe({
  
  rvalues$SessionTracking$log2db(
    action="access",
    param="home",
    desc=getOption("deployVersion")
  )
})
}
\keyword{datasets}
