% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymodules.R
\docType{package}
\name{tidymodules-package}
\alias{tidymodules}
\alias{tidymodules-package}
\title{tidymodules: A robust framework for developing shiny modules}
\description{
tidymodules offers a robust framework for developing shiny modules based on R6 classes which should facilitates inter-modules communication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Novartis/tidymodules}
  \item \url{http://opensource.nibr.com/tidymodules/}
  \item Report bugs at \url{https://github.com/Novartis/tidymodules/issues}
}

}
\author{
\strong{Maintainer}: Mustapha Larbaoui \email{mustapha.larbaoui@novartis.com}

Other contributors:
\itemize{
  \item Douglas Robinson \email{douglas.robinson@novartis.com} [contributor]
  \item Xiao Ni \email{xiao.ni@novartis.com} [contributor]
  \item David Granjon \email{david.granjon@novartis.com} [contributor]
  \item Stephen Eng \email{stefaneng13@gmail.com} [contributor]
  \item Marzieh Eslami Rasekh \email{marzie.rasekh@gmail.com} [contributor]
  \item Renan Sauteraud \email{rxs575@psu.edu} [contributor]
}

}
\keyword{internal}
