% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityModule.R
\docType{class}
\name{UtilityModule}
\alias{UtilityModule}
\title{Utility Module}
\format{\code{\link{R6Class}} object.}
\usage{
UtilityModule
}
\value{
Object of \code{\link{R6Class}}
}
\description{
This is a utility module used by some functions to retrieve the ModStore
}
\keyword{datasets}
