% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_module.R
\name{add_module}
\alias{add_module}
\title{Create a module}
\usage{
add_module(
  name,
  inherit = "TidyModule",
  path = getwd(),
  prefix = "tm",
  open = TRUE,
  dir_create = TRUE,
  export = FALSE
)
}
\arguments{
\item{name}{The class name of the module.}

\item{inherit}{Parent module class. Default is TidyModule.}

\item{path}{Where to created the file. Default is \code{getwd()}. The function will add \code{R} to the path if the sub-folder exists.}

\item{prefix}{filename prefix. Default is \code{tm}. Set to `NULL`` to disable.}

\item{open}{Should the file be opened?}

\item{dir_create}{Creates the directory if it doesn't exist, default is \code{TRUE}.}

\item{export}{Logical. Should the module be exported? Default is \code{FALSE}.}
}
\description{
This function creates a \code{{tm}} module class inside the current folder.
}
\note{
As a convention, this function will automatically capitalize the first character of the \code{name} argument.
}
