% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{callModules}
\alias{callModules}
\title{Call modules function}
\usage{
callModules()
}
\description{
This utility function call all modules initialized in the global session.
The global session is the session shared outside the server function of the application.
All the modules initialized in the global session can be called with this function in a single call.
The function take care of cloning and attaching them to the current user session.

Note that this function can only be called in the app server function at the moment.
We are working on supporting callModules within module server function for invoking nested modules.
}
