% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getMod}
\alias{getMod}
\title{Retrieve module from ModStore}
\usage{
getMod(id = 1, group = NULL)
}
\arguments{
\item{id}{Name or Id of the module}

\item{group}{Group name}
}
\description{
This utility function retrieve tidymodules from the central ModStore
using module namespace/id and/or group
}
\examples{

MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
MyModule$new("MyFirst")
MyModule$new("MySecond")
MyModule$new("MyThird",group = "B")

# MyFirst
getMod(1)
getMod("MyFirst")

# MySecond
getMod(2)

# MyThird
getMod(2)
getMod("B-MyThird")
getMod(1,group="B")

}
