% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%>>y\%}
\alias{\%>>y\%}
\title{Input port mapping function}
\value{
The left object
}
\description{
This pipe maps the left object (must be a reactive function or a reactivevalues object)
to the right module's input port defined by the number in the operator (y).
}
