% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{listModules}
\alias{listModules}
\title{List modules function}
\usage{
listModules(verbose = FALSE)
}
\description{
This function list the modules found in the global session

#' @param verbose Display module description
}
