% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{mkDoublePipe}
\alias{mkDoublePipe}
\title{mkDoublePipe: Pipe function generator}
\usage{
mkDoublePipe(l, r, f = TRUE, rev = FALSE)
}
\arguments{
\item{l}{Left hand side expression.}

\item{r}{Right hand side.}

\item{f}{Forward operartion. Boolean.}

\item{rev}{Reverse operation. Boolean.}
}
\description{
Create a pipe function for mapping module output to module input
}
\keyword{internal}
