% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{mkSinglePipe}
\alias{mkSinglePipe}
\title{mkSinglePipe: Pipe function generator}
\usage{
mkSinglePipe(p = NULL, f = TRUE, rev = FALSE)
}
\arguments{
\item{f}{Forward operartion. Boolean.}

\item{rev}{Reverse operation. Boolean.}
}
\description{
Create a pipe function for mapping a reactive expression/value to a module input
}
\keyword{internal}
