% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModStore.R
\name{ModStore}
\alias{ModStore}
\title{R6 Class Representing a ModStore}
\description{
This class is used to create a storage for tidymodules objects.
}
\details{
Manage applications, sessions and modules.
}
\examples{

## ------------------------------------------------
## Method `ModStore$new`
## ------------------------------------------------

MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
m <- MyModule$new()
s <- m$getStore()

## ------------------------------------------------
## Method `ModStore$isStored`
## ------------------------------------------------

MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
m <- MyModule$new()
s <- m$getStore()
s$isStored(m)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModStore-new}{\code{ModStore$new()}}
\item \href{#method-ModStore-isStored}{\code{ModStore$isStored()}}
\item \href{#method-ModStore-getGlobalSession}{\code{ModStore$getGlobalSession()}}
\item \href{#method-ModStore-getSession}{\code{ModStore$getSession()}}
\item \href{#method-ModStore-getSessions}{\code{ModStore$getSessions()}}
\item \href{#method-ModStore-getMods}{\code{ModStore$getMods()}}
\item \href{#method-ModStore-getEdges}{\code{ModStore$getEdges()}}
\item \href{#method-ModStore-addEdge}{\code{ModStore$addEdge()}}
\item \href{#method-ModStore-delEdges}{\code{ModStore$delEdges()}}
\item \href{#method-ModStore-addMod}{\code{ModStore$addMod()}}
\item \href{#method-ModStore-delMod}{\code{ModStore$delMod()}}
\item \href{#method-ModStore-print}{\code{ModStore$print()}}
\item \href{#method-ModStore-clone}{\code{ModStore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-new}{}}}
\subsection{Method \code{new()}}{
Create a new ModStore object.
Should be called once by the TidyModule class.
Not to be called directly outside TidyModule.
The ModStore object can be retrieved from any TidyModule object, see example below.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ModStore} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
m <- MyModule$new()
s <- m$getStore()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-isStored"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-isStored}{}}}
\subsection{Method \code{isStored()}}{
Check if a module is stored in the current session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$isStored(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
m <- MyModule$new()
s <- m$getStore()
s$isStored(m)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-getGlobalSession"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-getGlobalSession}{}}}
\subsection{Method \code{getGlobalSession()}}{
Retrieve the global session 'global_session'.
This is the session that exists outside the application server function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$getGlobalSession()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-getSession"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-getSession}{}}}
\subsection{Method \code{getSession()}}{
Retrieve a module session.
This could be the global session or a user session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$getSession(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-getSessions"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-getSessions}{}}}
\subsection{Method \code{getSessions()}}{
Retrieve all sessions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$getSessions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-getMods"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-getMods}{}}}
\subsection{Method \code{getMods()}}{
Retrieve all modules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$getMods(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-getEdges"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-getEdges}{}}}
\subsection{Method \code{getEdges()}}{
Retrieve modules connections.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$getEdges(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-addEdge"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-addEdge}{}}}
\subsection{Method \code{addEdge()}}{
Add modules connections into ModStore.
An edge is either a connection between a reactive object and a module
or between two modules.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$addEdge(from, to, mode = "direct", comment = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{list with three elements: m -> module, type -> input or output, port -> port Id.}

\item{\code{to}}{list with three elements: m -> module, type -> input or output, port -> port Id.}

\item{\code{mode}}{The type of edge, default to 'direct'.}

\item{\code{comment}}{Any additional comment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-delEdges"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-delEdges}{}}}
\subsection{Method \code{delEdges()}}{
Remove module edges
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$delEdges(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-addMod"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-addMod}{}}}
\subsection{Method \code{addMod()}}{
Add module into the ModStore.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$addMod(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-delMod"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-delMod}{}}}
\subsection{Method \code{delMod()}}{
Delete a module from the ModStore.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$delMod(m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{TidyModule object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-print"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-print}{}}}
\subsection{Method \code{print()}}{
Print the ModStore object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModStore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModStore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModStore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
