% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{multi_port_map}
\alias{multi_port_map}
\title{multi_port_map}
\usage{
multi_port_map(l_mod, r_mod, f = FALSE, t = NULL)
}
\arguments{
\item{l_mod}{Left module.}

\item{r_mod}{Right module.}

\item{f}{fast boolean. Used with the '>>' or '<<' signs to return the item at the opposite of the arrow.}

\item{t}{mapping type. Default to NULL. Could also be 'input' for mapping input to input.}
}
\description{
Pipe function for sequentially mapping left module outputs to
right module inputs
}
\keyword{internal}
