% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{multiPipeFunc}
\alias{multiPipeFunc}
\title{multiPipeFunc}
\usage{
multiPipeFunc(l_mod, r_mod, rev = FALSE, t = NULL)
}
\description{
Pipe function for sequentially mapping left module outputs to
right module inputs
}
\keyword{internal}
