% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{race_ports}
\alias{race_ports}
\title{Race ports function}
\usage{
race_ports(...)
}
\arguments{
\item{...}{List of racing ports}
}
\description{
This function collapse ports into a single port and make them race (i.e. always return the last one updated)
}
\examples{
\dontrun{
# Somewhere in the app...
MyModule <- R6::R6Class("MyModule", inherit = tidymodules::TidyModule)
MyModule$new("Mod1")
MyModule$new("Mod2")
MyModule$new("Mod3")

# Must be in the server code and after calling the modules!
callModules()
observe({
  race_ports(
    mod(1)$getOutput(1),
    mod(2)$getOutput(1)
  ) \%>1\% mod(3)
})
}

}
