% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Store.R
\name{Store}
\alias{Store}
\title{This TidyModule is used to explore the content of the ModStore.}
\description{
Store is a TidyModule that can be used in your application to list existing applications, sessions and display your session's modules and edges.
}
\details{
Should be initialized and injected in your application.
}
\section{Super class}{
\code{\link[tidymodules:TidyModule]{tidymodules::TidyModule}} -> \code{Store}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Store-ui}{\code{Store$ui()}}
\item \href{#method-Store-server}{\code{Store$server()}}
\item \href{#method-Store-clone}{\code{Store$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="addInputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-addInputPort'><code>tidymodules::TidyModule$addInputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="addOutputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-addOutputPort'><code>tidymodules::TidyModule$addOutputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="assignPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-assignPort'><code>tidymodules::TidyModule$assignPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="callModule"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-callModule'><code>tidymodules::TidyModule$callModule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="countInputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-countInputPort'><code>tidymodules::TidyModule$countInputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="countOutputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-countOutputPort'><code>tidymodules::TidyModule$countOutputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="deepClone"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-deepClone'><code>tidymodules::TidyModule$deepClone()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="definePort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-definePort'><code>tidymodules::TidyModule$definePort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="doServer"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-doServer'><code>tidymodules::TidyModule$doServer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="execInput"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-execInput'><code>tidymodules::TidyModule$execInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="execOutput"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-execOutput'><code>tidymodules::TidyModule$execOutput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getGlobalSession"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getGlobalSession'><code>tidymodules::TidyModule$getGlobalSession()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getInput"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getInput'><code>tidymodules::TidyModule$getInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getInputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getInputPort'><code>tidymodules::TidyModule$getInputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getInputPorts"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getInputPorts'><code>tidymodules::TidyModule$getInputPorts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getOutput"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getOutput'><code>tidymodules::TidyModule$getOutput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getOutputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getOutputPort'><code>tidymodules::TidyModule$getOutputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getOutputPorts"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getOutputPorts'><code>tidymodules::TidyModule$getOutputPorts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getPortDef"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getPortDef'><code>tidymodules::TidyModule$getPortDef()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getPortName"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getPortName'><code>tidymodules::TidyModule$getPortName()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getSession"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getSession'><code>tidymodules::TidyModule$getSession()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getSessionId"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getSessionId'><code>tidymodules::TidyModule$getSessionId()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getShinyInput"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getShinyInput'><code>tidymodules::TidyModule$getShinyInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getShinyOutput"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getShinyOutput'><code>tidymodules::TidyModule$getShinyOutput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getShinySession"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getShinySession'><code>tidymodules::TidyModule$getShinySession()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="getStore"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-getStore'><code>tidymodules::TidyModule$getStore()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="initialize"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-initialize'><code>tidymodules::TidyModule$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="iport"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-iport'><code>tidymodules::TidyModule$iport()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="isGlobal"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-isGlobal'><code>tidymodules::TidyModule$isGlobal()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="isStored"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-isStored'><code>tidymodules::TidyModule$isStored()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="ns"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-ns'><code>tidymodules::TidyModule$ns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="oport"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-oport'><code>tidymodules::TidyModule$oport()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="print"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-print'><code>tidymodules::TidyModule$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="render"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-render'><code>tidymodules::TidyModule$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="reset"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-reset'><code>tidymodules::TidyModule$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="updateInputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-updateInputPort'><code>tidymodules::TidyModule$updateInputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="updateInputPorts"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-updateInputPorts'><code>tidymodules::TidyModule$updateInputPorts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="updateOutputPort"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-updateOutputPort'><code>tidymodules::TidyModule$updateOutputPort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="updateOutputPorts"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-updateOutputPorts'><code>tidymodules::TidyModule$updateOutputPorts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="use"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-use'><code>tidymodules::TidyModule$use()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tidymodules" data-topic="TidyModule" data-id="view"><a href='../../tidymodules/html/TidyModule.html#method-TidyModule-view'><code>tidymodules::TidyModule$view()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-ui"></a>}}
\if{latex}{\out{\hypertarget{method-Store-ui}{}}}
\subsection{Method \code{ui()}}{
Store's ui function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$ui()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
UI elements.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-server"></a>}}
\if{latex}{\out{\hypertarget{method-Store-server}{}}}
\subsection{Method \code{server()}}{
Store's server function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Shiny input.}

\item{\code{output}}{Shiny output}

\item{\code{session}}{Shiny session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Store-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Store-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Store$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
